requirements.txt
# --- Core Backend Framework ---
fastapi==0.115.12
uvicorn==0.34.0
gunicorn==23.0.0

# --- Async & I/O ---
aiofiles==24.1.0
aioboto3==14.3.0
httpx==0.28.1
aiokafka==0.12.0

# --- AWS / GCP Integration ---
gcloud-aio-storage==9.4.0
python-dotenv==1.1.0

# --- ML / DL Libraries ---
numpy==2.2.5
pandas==2.2.3
scikit-learn==1.6.1
scipy==1.15.2
joblib==1.4.2
torch==2.7.0
torchvision==0.22.0
torchaudio==2.7.0
thop==0.1.1.post2209072238  # FLOPs 계산용

# --- Visualization (필요 없으면 제거 가능) ---
matplotlib==3.10.1
seaborn==0.13.2

# --- Image Processing ---
opencv-contrib-python==4.11.0.86
Pillow==11.2.1  # pillow와 중복, 하나만 유지

# --- Logging / Utils ---
loguru==0.7.3
tqdm==4.67.1
cachetools==5.3.3
python-dateutil==2.9.0.post0
typing_extensions==4.13.2

# --- Data Format / Parsing ---
PyYAML==6.0.2
packaging==25.0
pyparsing==3.2.3

# --- Web ---
Jinja2==3.1.6
MarkupSafe==3.0.2
requests==2.32.4
urllib3==2.4.0
idna==3.10
certifi==2025.1.31
charset-normalizer==3.4.1

# --- Math / Symbolic ---
sympy==1.13.3
mpmath==1.3.0

# --- Graph / Utils ---
networkx==3.4.2
fsspec==2025.3.2
filelock==3.18.0
threadpoolctl==3.6.0
six==1.17.0
tzdata==2025.2

# --- Font / Plotting Internals ---
fonttools==4.57.0
kiwisolver==1.4.8
cycler==0.12.1

# --- Model Validation ---
pydantic==2.11.3

# --- Caching ---
redis==6.2.0

# --- Setup Tool (빌드시만 필요, 실행시 제거 가능) ---
setuptools==79.0.1

# --- CI/CD ---
google-cloud-secret-manager==2.24.0

# --- Monitoring ---
prometheus_fastapi_instrumentator